gym_icesi/requirements.txt
Django==5.2.8
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1

